requirements.txt
cryptography
mysql-connector-python
watchdog